fix: rename sw-trigger to sw-bootstrap
