+++ packages/rax-plugin-pwa/src/templates/sw.js
